+++ NEWS.md
@@ -1,6 +1,7 @@
 Version 2.0.5
 =============
 * num.RR and num.lv with lvCor can now be combined
+* lvCor can be used with multiple (iid) terms
 
 # Bugfixes
 * Bug fixed in VP for mixed effects formula